// ui/src/generated/overlay-placement-plan.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Disposition = "place" | "defer";
export type Left = number;
export type Top = number;
export type Width = number;
export type Height = number;
export type MonitorName = string | null;
export type Anchor = string | null;
export type Reason = string;
export type Density = number;

export interface OverlayPlacementPlan {
  disposition: Disposition;
  target: Rect;
  monitor_name: MonitorName;
  anchor: Anchor;
  reason: Reason;
  density: Density;
}
export interface Rect {
  left: Left;
  top: Top;
  width: Width;
  height: Height;
}