// ui/src/generated/overlay-placement-request.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type PlacementIntent = "conversation" | "proactive";
export type Left = number;
export type Top = number;
export type Width = number;
export type Height = number;
export type Visible = boolean;
export type X = number;
export type Y = number;
/**
 * @minItems 1
 * @maxItems 8
 */
export type Monitors =
  | [MonitorLayout]
  | [MonitorLayout, MonitorLayout]
  | [MonitorLayout, MonitorLayout, MonitorLayout]
  | [MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout]
  | [MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout]
  | [MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout]
  | [MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout, MonitorLayout]
  | [
      MonitorLayout,
      MonitorLayout,
      MonitorLayout,
      MonitorLayout,
      MonitorLayout,
      MonitorLayout,
      MonitorLayout,
      MonitorLayout
    ];
export type Name = string;
export type Left1 = number;
export type Top1 = number;
export type Width1 = number;
export type Height1 = number;

export interface OverlayPlacementRequest {
  intent: PlacementIntent;
  overlay: OverlayGeometry;
  pointer: Point;
  monitors: Monitors;
}
export interface OverlayGeometry {
  left: Left;
  top: Top;
  width: Width;
  height: Height;
  visible: Visible;
}
export interface Point {
  x: X;
  y: Y;
}
export interface MonitorLayout {
  name: Name;
  bounds: Rect;
  work_area: Rect;
}
export interface Rect {
  left: Left1;
  top: Top1;
  width: Width1;
  height: Height1;
}